Finilized the project and added contact form functionality
- add Navbar component and integrate it into Home and Contact pages
- implement ThemeToggle component and update imports
- enhance dashboard Navbar with contact link
- add edit functionality to ShortUrlCard component
- refactor DashboardContext to use useCallback for fetchShortUrls
- create separate useDashboardContext hook for better context management
- implement Contact page with form handling and validation
- remove unused static pages (Help, Policy, Terms)
- update router to remove references to deleted static pages
- add dompurify for sanitizing user input in Contact form
- update package.json and package-lock.json with

+++ backend/src/user/user.service.ts
@@ -211,6 +211,14 @@ export class UserService {
         throw new Error('Failed to delete user bookmarks');
       }
 
+      const shorturls = await this.prisma.shortUrl.deleteMany({
+        where: { createdById: userId },
+      });
+
+      if (!shorturls) {
+        throw new Error('Failed to delete user short URLs');
+      }
+
       const emails = await this.prisma.email.deleteMany({
         where: { userId },
       });
